Port pcc to Mac OS X.  The main test suite passes with no compile failures and 5 run failures (which look like they're due to float output differences).  The std, pcre, curl, and sqlite extensions have been built.
The main thing I had to deal with when porting was the different
linker.  On the mac, we create .dylib files.  The "install_name"
matters, so we have to pass it explicitly (see Makefile.config.in).
The linker doesn't like a space between -L and the directory, so I
whacked run-command to smoosh them together.

There were some slightly different header files and paths, and I took
out uname() instead of making it work right.

No web backends or library support has been tested.