add support for extensions written in PHP. these are essentially the same as our normal compiled PHP libraries, but now a library can be initialized upon load by running an include file specified during compile.
this allows top level code to be run in the library without the user
specifically include()'ing one of the library component files which
is otherwise necessary.

a basic PDO extension is the first PHP extension. it provides the
PDO database classes for mysql and sqlite currently. this code is
from a GPL PDO PHP 4 emulation layer, but will be modified to offer real 
PHP 5 features like class const, etc.

the mysql and sqlite PDO classes are enabled during configure if either of
those normal extensions are enabled.