move builtin-interfaces.scm to builtin-classes.scm and define Exception there
add define-builtin-php-class which allows the definition of a class
that is not wiped during a runtime reset. add a new class flag "pcc_builtin"
that can be used in php land to define these classes, and make our current
builtin classes use it. some rearranging of the runtime-reset calls to fit
in with this scheme